requirements.txt
python-multipart
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
jinja2
psycopg2-binary
python-dotenv
passlib[bcrypt]
pydantic[email]
pydantic-settings